finish tests

+++ src/main/java/com/skillbox/socialnet/controller/ProfileController.java
@@ -1,29 +1,26 @@
 package com.skillbox.socialnet.controller;
 
 import com.skillbox.socialnet.model.RQ.PostChangeRQ;
-import com.skillbox.socialnet.model.RQ.UserSearchRQ;
 import com.skillbox.socialnet.model.RQ.UserChangeRQ;
-import com.skillbox.socialnet.model.RS.DefaultRS;
+import com.skillbox.socialnet.model.RQ.UserSearchRQ;
 import com.skillbox.socialnet.model.RS.GeneralListResponse;
 import com.skillbox.socialnet.model.RS.GeneralResponse;
 import com.skillbox.socialnet.model.dto.MessageOkDTO;
 import com.skillbox.socialnet.model.dto.PostDTO;
 import com.skillbox.socialnet.model.dto.UserDTO;
-import com.skillbox.socialnet.model.mapper.DefaultRSMapper;
-import com.skillbox.socialnet.security.JwtProvider;
+import com.skillbox.socialnet.service.AuthService;
 import com.skillbox.socialnet.service.PostService;
 import com.skillbox.socialnet.service.UserService;
 import com.skillbox.socialnet.util.ElementPageable;
 import com.skillbox.socialnet.util.anotation.MethodLog;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.util.Calendar;
-import java.util.List;
 
 @CrossOrigin
 @RestController
@@ -33,10 +30,14 @@ public class ProfileController {
 
     private final UserService userService;
     private final PostService postService;
+    private final AuthService authService;
 
     @MethodLog
     @GetMapping("/me")
     public ResponseEntity<GeneralResponse<UserDTO>> getUser() {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.getUser()));
     }
 
@@ -45,11 +46,17 @@ public class ProfileController {
     public ResponseEntity<GeneralResponse<UserDTO>> editUser(
             @RequestBody @Valid UserChangeRQ userChangeRQ,
             HttpServletRequest request) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.editUser(userChangeRQ)));
     }
 
     @DeleteMapping("/me")
     public ResponseEntity<GeneralResponse<String>> deleteUser() {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.deleteUser()));
     }
 
@@ -69,6 +76,9 @@ public class ProfileController {
             @PathVariable int id,
             @RequestParam(name = "publish_date", defaultValue = "0") long publishDate,
             @RequestBody @Valid PostChangeRQ postChangeRQ) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(postService.addPostToUserWall(id, publishDate, postChangeRQ)));
     }
 
@@ -95,11 +105,17 @@ public class ProfileController {
 
     @PutMapping("/block/{id}")
     public ResponseEntity<GeneralResponse<MessageOkDTO>> blockUser(@PathVariable int id) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.blockUser(id)));
     }
 
     @DeleteMapping("/block/{id}")
     public ResponseEntity<GeneralResponse<MessageOkDTO>> unblockUser(@PathVariable int id) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.unblockUser(id)));
     }
